refactor FaceColoring
